backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.110.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.9.1
pydantic_core==2.18.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
google-generativeai==0.5.0